+++ src/resources/payments.ts
@@ -6,8 +6,8 @@ import { IRazorTransfer, IRazorTransferId } from "./routes";
 import { IRazorBankTransferId, IRazorUpiTransferId } from "./virtual-accounts";
 
 export interface IRazorPayment {
-    entity?: string;
-    amount: string;
+    entity?: 'payment';
+    amount: number;
     currency: string;
     base_amount?: string;
     base_currency?: string;
@@ -22,6 +22,13 @@ export interface IRazorPayment {
     contact?: string;
     tax?: number;
     notes?: { [key: string]: string; };
+    invoice_id?: string | null;
+    card_id?: string | null;
+    bank?: string | null;
+    wallet?: string | null;
+    vpa?: string | null;
+    customer_id?: string | null;
+    token_id?: string | null;
 }
 
 export interface IRazorPaymentId extends IRazorPayment {